Fix: handling of 'from' and scroll to top
There was an issue when changing the 'sort by' value while navigating through different pages. The 'from' value would not update back to 0 and cause you to stay on the same page or default to page 2. Scrolling back to the top would break as well and not always work. By moving the handleFrom and scrollToTop functions into their own useEffect, the pagination and scrolling now work correctly.